+++ src/components/EmployeeForm.js
@@ -1,0 +1,72 @@
+import React, { Component } from 'react';
+import { View, Text, Picker } from 'react-native';
+import { connect } from 'react-redux';
+import PropTypes from 'prop-types';
+
+import { CardSection, Input } from './common';
+import { employeeUpdate } from '../actions';
+
+class EmployeeForm extends Component {
+  render() {
+    return (
+      <View>
+        <CardSection>
+          <Input
+            label="Name"
+            placeholder="Jane"
+            value={this.props.name}
+            onChangeText={value => this.props.employeeUpdate({ prop: 'name', value })}
+          />
+        </CardSection>
+
+        <CardSection>
+          <Input
+            label="Phone"
+            placeholder="555-555-5555"
+            value={this.props.phone}
+            onChangeText={value => this.props.employeeUpdate({ prop: 'phone', value })}
+          />
+        </CardSection>
+
+        <CardSection style={{ flexDirection: 'column' }}>
+          <Text style={styles.pickerTextStyle}>Shift</Text>
+          <Picker
+            selectedValue={this.props.shift}
+            onValueChange={value => this.props.employeeUpdate({ prop: 'shift', value })}
+          >
+            <Picker.Item label="Monday" value="Monday" />
+            <Picker.Item label="Tuesday" value="Tuesday" />
+            <Picker.Item label="Wednesday" value="Wednesday" />
+            <Picker.Item label="Thursday" value="Thursday" />
+            <Picker.Item label="Friday" value="Friday" />
+            <Picker.Item label="Saturday" value="Saturday" />
+            <Picker.Item label="Sunday" value="Sunday" />
+          </Picker>
+        </CardSection>
+      </View>
+    );
+  }
+}
+
+EmployeeForm.propTypes = {
+  name: PropTypes.string,
+  phone: PropTypes.string,
+  shift: PropTypes.string,
+  employeeUpdate: PropTypes.func,
+};
+
+const styles = {
+  pickerTextStyle: {
+    fontSize: 18,
+    paddingLeft: 20,
+    paddingTop: 7,
+  }
+};
+
+const mapStateToProps = state => {
+  const { name, phone, shift } = state.employeeForm;
+
+  return { name, phone, shift };
+};
+
+export default connect(mapStateToProps, { employeeUpdate })(EmployeeForm);
