+++ src/screens/EmployeeCreateScreen.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
-import { Picker, Text } from 'react-native';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
-import { Card, CardSection, Input, Button } from '../components/common';
-import { employeeUpdate, employeeCreate } from '../actions';
+import { Card, CardSection, Button } from '../components/common';
+import { employeeCreate } from '../actions';
+import EmployeeForm from '../components/EmployeeForm';
 
-class EmployeeCreate extends Component {
+class EmployeeCreateScreen extends Component {
   static navigationOptions = {
     headerTitle: 'Add Employee'
   };
@@ -22,39 +22,7 @@ class EmployeeCreate extends Component {
     
     return (
       <Card>
-        <CardSection>
-          <Input
-            label="Name"
-            placeholder="Jane"
-            value={this.props.name}
-            onChangeText={value => this.props.employeeUpdate({ prop: 'name', value })}
-          />
-        </CardSection>
-
-        <CardSection>
-          <Input
-            label="Phone"
-            placeholder="555-555-5555"
-            value={this.props.phone}
-            onChangeText={value => this.props.employeeUpdate({ prop: 'phone', value })}
-          />
-        </CardSection>
-
-        <CardSection style={{ flexDirection: 'column' }}>
-          <Text style={styles.pickerTextStyle}>Shift</Text>
-          <Picker
-            selectedValue={this.props.shift}
-            onValueChange={value => this.props.employeeUpdate({ prop: 'shift', value })}
-          >
-            <Picker.Item label="Monday" value="Monday" />
-            <Picker.Item label="Tuesday" value="Tuesday" />
-            <Picker.Item label="Wednesday" value="Wednesday" />
-            <Picker.Item label="Thursday" value="Thursday" />
-            <Picker.Item label="Friday" value="Friday" />
-            <Picker.Item label="Saturday" value="Saturday" />
-            <Picker.Item label="Sunday" value="Sunday" />
-          </Picker>
-        </CardSection>
+        <EmployeeForm {...this.props} />
 
         <CardSection>
           <Button onPress={this.onButtonPress.bind(this)}>
@@ -66,29 +34,18 @@ class EmployeeCreate extends Component {
   }
 }
 
-EmployeeCreate.propTypes = {
+EmployeeCreateScreen.propTypes = {
   name: PropTypes.string,
   phone: PropTypes.string,
-  employeeUpdate: PropTypes.func,
   employeeCreate: PropTypes.func,
   shift: PropTypes.string,
   navigation: PropTypes.object,
 };
 
-const styles = {
-  pickerTextStyle: {
-    fontSize: 18,
-    paddingLeft: 20,
-    paddingTop: 7,
-  }
-};
-
-const mapStateToProps = (state) => {
+const mapStateToProps = state => {
   const { name, phone, shift } = state.employeeForm;
 
   return { name, phone, shift };
 };
 
-export default connect(mapStateToProps, {
-  employeeUpdate, employeeCreate
-})(EmployeeCreate);
+export default connect(mapStateToProps, { employeeCreate })(EmployeeCreateScreen);
